requirements.txt
discord.py==1.7.3
discord_components==2.1.2
